[Feat] MainViewController 통합
WalkMainViewController삭제
